requirements.txt
streamlit
pandas
numpy
pdfplumber
Pillow
Wand
stqdm
pdf2jpg
pdf2image